backend/requirements.txt
fastapi==0.116.1
google-genai==1.69.0
httpx==0.28.1
pydantic-settings==2.11.0
python-dotenv==1.1.1
supabase==2.28.3
uvicorn[standard]==0.38.0
